Widen Label from int to long for 64-bit ASCII-encoded syslib labels
Change Label field, constructor, delegate, and Evaluate method from
int to long. Update compiler code generation to emit L-suffixed long
literals. Remove 16-bit label validation limit. Existing syslib labels
(1000-1999) continue to work via implicit int→long conversion.

This enables the i# syslib to use labels like DIVIDE16 (0x4449564944453136)
encoded as ASCII bytes packed into 64-bit values.

+++ intercal.runtime/twisty.cs
@@ -42,9 +42,9 @@ namespace INTERCAL.Runtime
         //now true means "terminate" ("you've been forgotten") and 
         //false means "continue" ("you've been resumed");
         public bool Result;
-        public int Label;
+        public long Label;
 
-        public ExecutionFrame(ExecutionContext context, IntercalThreadProc proc, int label)
+        public ExecutionFrame(ExecutionContext context, IntercalThreadProc proc, long label)
         {
             ExecutionContext = context;
             Proc = proc;
@@ -105,7 +105,7 @@ namespace INTERCAL.Runtime
         protected Exception CurrentException { get; set; }
         protected Stack<ExecutionFrame> NextingStack = new Stack<ExecutionFrame>();
 
-        protected delegate bool StartProc(IntercalThreadProc proc, int label);
+        protected delegate bool StartProc(IntercalThreadProc proc, long label);
         public void Resume(uint depth)
         {
             //depth must be zero.  We depend on the compiler to ensure that
